+++ lib/components/menu/desktop_title.dart
@@ -18,38 +18,41 @@ class BodyTitle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Row(
-      mainAxisSize: MainAxisSize.max,
-      children: [
-        // ANIMATED ICON -------------------------------------------------------
-        GetBuilder(
-          init: router,
-          builder: (_) => AnimatedSize(
-            duration: animDurationShort,
-            child: AnimatedSwitcher(
+    return GestureDetector(
+      onTap: () => router.pop(),
+      child: Row(
+        mainAxisSize: MainAxisSize.max,
+        children: [
+          // ANIMATED ICON -------------------------------------------------------
+          GetBuilder(
+            init: router,
+            builder: (_) => AnimatedSize(
               duration: animDurationShort,
-              child: router.atHome
-                  ? const SizedBox()
-                  : Padding(
-                      padding: EdgeInsets.only(right: XLayout.paddingXS),
-                      child: XeppelinLogo(
-                        size: 2 * XLayout.paddingL,
-                      )),
+              child: AnimatedSwitcher(
+                duration: animDurationShort,
+                child: router.atHome
+                    ? const SizedBox()
+                    : Padding(
+                        padding: EdgeInsets.only(right: XLayout.paddingXS),
+                        child: XeppelinLogo(
+                          size: 2 * XLayout.paddingL,
+                        )),
+              ),
             ),
           ),
-        ),
 
-        // TEXT ----------------------------------------------------------------
-        Expanded(
-          child: FittedBox(
-            fit: BoxFit.fitWidth,
-            child: Text(
-              "Xeppelin",
-              style: context.textTheme.displayLarge,
+          // TEXT ----------------------------------------------------------------
+          Expanded(
+            child: FittedBox(
+              fit: BoxFit.fitWidth,
+              child: Text(
+                "Xeppelin",
+                style: context.textTheme.displayLarge,
+              ),
             ),
           ),
-        ),
-      ],
+        ],
+      ),
     );
   }
 }