+++ lib/components/projects/projects_menu.dart
@@ -1,9 +1,7 @@
 import "package:flutter/material.dart";
-import "package:get/get.dart";
 import "package:x_containers/x_containers.dart";
 
-import "../../utils/exports.dart";
-import "../../widgets/images/xeppelin_logo.dart";
+import "../menu/desktop_title.dart";
 import "projects_menu_list.dart";
 
 /// The half of the body containing the title and the menu.
@@ -24,21 +22,22 @@ class ProjectsMenu extends StatelessWidget {
       child: Column(
         children: [
           // TITLE ---------------------------------------------------------------
-          XInkContainer(
-            child: Row(
-              children: [
-                XeppelinLogo(
-                  size: 2 * XLayout.paddingL,
-                ),
-                XLayout.horizontalXS,
-                Text(
-                  "Xeppelin",
-                  style: context.textTheme.displayLarge,
-                ),
-              ],
-            ),
-            onTap: () => router.pop(),
-          ),
+          // XInkContainer(
+          //   child: Row(
+          //     children: [
+          //       XeppelinLogo(
+          //         size: 2 * XLayout.paddingL,
+          //       ),
+          //       XLayout.horizontalXS,
+          //       Text(
+          //         "Xeppelin",
+          //         style: context.textTheme.displayLarge,
+          //       ),
+          //     ],
+          //   ),
+          //   onTap: () => router.pop(),
+          // ),
+          const BodyTitle(),
 
           // SEPARATOR
           XLayout.verticalM,